tests: update igraph_avg_nearest_neighbor_degree() example
