+++ examples/simple/igraph_avg_nearest_neighbor_degree.out
@@ -1,0 +1,4 @@
+knn: 4.3125 5.77778 6.6 7.66667 7.66667 6.25 6.25 10.25 11.8 13.5 7.66667 16 11 11.6 14.5 14.5 4 12.5 14.5 14 14.5 12.5 14.5 8 4.33333 4.66667 10.5 8.75 11 9 10.75 9 5.08333 3.82353
+knn(k): 16 12.4091 8.22222 8.54167 10.4667 8.33333 NaN NaN 5.77778 6.6 NaN 5.08333 NaN NaN NaN 4.3125 3.82353
+knn: 3.45833 4.28205 5.4346 5.55634 4 3.42373 3.52991 8.64198 11.1104 14.2192 4.73171 16 8.32558 11.6143 14.5269 14.5258 4 11.625 14.5248 14.8953 14.5234 11.7222 14.5225 8.05085 4.34921 4.67935 10.5489 8.81395 11.2892 9.30741 12.0664 8.31319 5.35303 4.05771
+knn(k): 16 12.4884 7.07042 8.37055 9.71906 7.53953 NaN NaN 4.28205 5.4346 NaN 5.35303 NaN NaN NaN 3.45833 4.05771
